--- src/components/ButtonGroup.tsx
@@ -1,43 +1,0 @@
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
-import ButtonGroup from "@mui/material/ButtonGroup";
-
-const buttons = [
-  <Button key="one">One</Button>,
-  <Button key="two">Two</Button>,
-  <Button key="three">Three</Button>,
-];
-
-export default function GroupOrientation() {
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        "& > *": {
-          m: 1,
-        },
-      }}
-    >
-      <ButtonGroup
-        orientation="vertical"
-        aria-label="vertical outlined button group"
-      >
-        {buttons}
-      </ButtonGroup>
-      <ButtonGroup
-        orientation="vertical"
-        aria-label="vertical contained button group"
-        variant="contained"
-      >
-        {buttons}
-      </ButtonGroup>
-      <ButtonGroup
-        orientation="vertical"
-        aria-label="vertical contained button group"
-        variant="text"
-      >
-        {buttons}
-      </ButtonGroup>
-    </Box>
-  );
-}
