+++ src/pages/Books.tsx
@@ -1,4 +1,12 @@
-import { Card, CardContent, CardMedia, Typography } from "@mui/material";
+import {
+  Box,
+  Button,
+  ButtonGroup,
+  Card,
+  CardContent,
+  CardMedia,
+  Typography,
+} from "@mui/material";
 import { useState } from "react";
 import { Link } from "react-router-dom";
 import ShopButton from "../components/ShopButton";
@@ -6,9 +14,18 @@ import { mockedProducts } from "../data";
 
 function Books() {
   const [books] = useState(mockedProducts);
-  const [filteredBooks, setFilteredBooks] = useState(books);
-
-  const categories = [...new Set(books.map((book) => book.category))];
+  const [categoryFilter, setCategoryFilter] = useState("");
+  const categories = [...new Set(mockedProducts.map((book) => book.category))];
+  const buttons = [
+    <Button key={-1} onClick={() => setCategoryFilter("")}>
+      Alla böcker
+    </Button>,
+    categories.map((c) => (
+      <Button key={categories.indexOf(c)} onClick={() => setCategoryFilter(c)}>
+        {c}
+      </Button>
+    )),
+  ];
 
   return (
     <main
@@ -27,11 +44,23 @@ function Books() {
           padding: ".4rem",
         }}
       >
-        {categories.map((category) => (
-          <div key={categories.indexOf(category)}>
-            <h4 style={{ margin: 0, padding: ".5rem" }}>{category}</h4>
-          </div>
-        ))}
+        <Box
+          sx={{
+            display: "flex",
+            "& > *": {
+              m: 1,
+            },
+          }}
+        >
+          <ButtonGroup
+            orientation="vertical"
+            aria-label="vertical contained button group"
+            variant="text"
+          >
+            {buttons}
+          </ButtonGroup>
+        </Box>
+        {/* <ButtonGroupOrientation /> */}
       </div>
       <div
         style={{
@@ -41,46 +70,50 @@ function Books() {
           flex: 7,
         }}
       >
-        {books.map((book) => (
-          <div
-            key={book.id}
-            style={{
-              padding: ".5rem",
-            }}
-          >
-            <Card
-              sx={{
-                maxWidth: "12rem",
-                minHeight: "30rem",
-                display: "flex",
-                flexDirection: "column",
+        {books
+          .filter((b) =>
+            b.category.toLowerCase().includes(categoryFilter.toLowerCase())
+          )
+          .map((book) => (
+            <div
+              key={book.id}
+              style={{
+                padding: ".5rem",
               }}
             >
-              <Link to={`/book/${book.id}`}>
-                <CardMedia
-                  component="img"
-                  height="300"
-                  image={book.image}
-                  alt={book.title + " cover"}
-                />
-              </Link>
-              <CardContent>
-                <Typography gutterBottom variant="body2">
-                  {book.title}
-                </Typography>
-                <Typography
-                  variant="body2"
-                  color="text.secondary"
-                  style={{ paddingBottom: ".3rem" }}
-                >
-                  {book.author}
-                </Typography>
+              <Card
+                sx={{
+                  maxWidth: "12rem",
+                  minHeight: "30rem",
+                  display: "flex",
+                  flexDirection: "column",
+                }}
+              >
+                <Link to={`/book/${book.id}`}>
+                  <CardMedia
+                    component="img"
+                    height="300"
+                    image={book.image}
+                    alt={book.title + " cover"}
+                  />
+                </Link>
+                <CardContent>
+                  <Typography gutterBottom variant="body2">
+                    {book.title}
+                  </Typography>
+                  <Typography
+                    variant="body2"
+                    color="text.secondary"
+                    style={{ paddingBottom: ".3rem" }}
+                  >
+                    {book.author}
+                  </Typography>
 
-                <ShopButton product={book}>{book.price} kr</ShopButton>
-              </CardContent>
-            </Card>
-          </div>
-        ))}
+                  <ShopButton product={book}>{book.price} kr</ShopButton>
+                </CardContent>
+              </Card>
+            </div>
+          ))}
       </div>
     </main>
   );
